+++ packages/app/src/app/contract/[id]/[tab]/layout.tsx
@@ -1,0 +1,5 @@
+import { ContractLayout } from '~/systems/Contract/components/ContractLayout';
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return <ContractLayout>{children}</ContractLayout>;
+}
