--- packages/app/src/app/contract/[id]/code/loading.tsx
@@ -1,5 +1,0 @@
-import { CodeBlock } from '~/systems/Core/components/CodeBlock/CodeBlock';
-
-export default function Loader() {
-  return <CodeBlock isLoading value="" />;
-}
