--- packages/app/src/app/contract/[id]/layout.tsx
@@ -1,11 +1,0 @@
-import { ContractLayout } from '~/systems/Contract/components/ContractLayout';
-
-export default function Layout({
-  children,
-  params: { id },
-}: {
-  children: React.ReactNode;
-  params: { id: string };
-}) {
-  return <ContractLayout id={id}>{children}</ContractLayout>;
-}
