--- packages/app/src/app/contract/[id]/loading.tsx
@@ -1,5 +1,0 @@
-import { AssetsLoader } from '~/systems/Asset/components/AssetsLoader';
-
-export default function Loading() {
-  return <AssetsLoader />;
-}
