+++ packages/app/src/app/contract/[id]/page.tsx
@@ -1,0 +1,11 @@
+import { redirect } from 'next/navigation';
+
+type ContractProps = {
+  params: {
+    id: string;
+  };
+};
+
+export default function Contract({ params: { id } }: ContractProps) {
+  redirect(`/contract/${id}/assets`);
+}
